rate-limiter-demo/pyproject.toml
[project]
name = "rate-limiter-demo"
version = "0.1.0"
description = "Rate limiter demo: Token Bucket, Leaking Bucket, Fixed Window, Sliding Window Log, Sliding Window Counter"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "redis>=5.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
